personas/pyproject.toml
[project]
name = "personas"
version = "0.1.0"
description = "AI-generated persona headshots and Claude reviewer commands for JWT parsing examples"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "Aidan Pace", email = "apace@defrecord.com"}
]
license = {text = "MIT"}
dependencies = [
    "orgparse",
    "pillow",
    "openai",
    "google-genai",
    "ruff>=0.11.7",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["scripts"]

[tool.hatch.metadata]
allow-direct-references = true

[project.scripts]
generate-headshots = "scripts.generate_headshots:main"
extract-persona = "scripts.extract_persona:main"

[project.optional-dependencies]
dev = [
    "ruff",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # isort
    "W",   # pycodestyle warnings
    "B",   # flake8-bugbear
]
ignore = ["E203"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
